requirements.txt
Django>=4.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
dj-database-url
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
pypdf
langchain-postgres
pgvector
langchain-huggingface
huggingface_hub
sentence-transformers
faiss-cpu
python-logging-loki

